added update for video games

+++ web/Inventory/console.php
@@ -1,12 +1,42 @@
 <?php
 session_start();
 include 'database.php';
+
+$_SESSION['table'] = 'video_games';
+
+if (!isset($_SESSION['id'])){
+	header("Location: signin.php");
+	die(); 
+}
+if (isset($_POST['update'])) {
+	$_SESSION['consoleid'] = $_POST['update'];
+	header("Location: update.php");
+	die();
+}
 if (isset($_POST)) {
 	if (isset($_POST['delete'])) {
 		$db->query('DELETE FROM video_games WHERE id =' . $_POST['delete']);
 	}
-	elseif (isset($_POST['update'])) {
-		
+elseif (isset($_POST['updateConsole'])) {
+		$vgT = $_POST['vgTitle'];
+		$vgST = $_POST['vgSubTitle'];
+		$vgR = $_POST['vgRating'];
+		$vgG = $_POST['vgGenre'];
+		$console = $_POST['console'];
+		$user = $_SESSION['id'];
+
+		$sql = 'UPDATE video_games 
+				SET title = :title, subtitle = :subtitle, rating = :rating, genre = :genre, console = :console
+				WHERE id = ' . $_SESSION['consoleid'];
+
+		$prep = $db->prepare($sql);
+		$prep->bindParam(':title', $vgT);
+		$prep->bindParam(':subtitle', $vgST);
+		$prep->bindParam(':rating', $vgR);
+		$prep->bindParam(':genre', $vgG);
+		$prep->bindParam(':console', $console);
+
+		$prep->execute();
 	}
 	else {
 		$vgT = $_POST['vgTitle'];
@@ -56,7 +86,7 @@ if (isset($_POST)) {
 <?php
 	foreach ($db->query('SELECT v.id, v.title, v.subtitle, g.genre, c.console, r.rating FROM video_games v
 	INNER JOIN rating r ON v.rating = r.id INNER JOIN console c ON v.console = c.id 
-	INNER JOIN genre g ON v.genre = g.id WHERE v.user_id = 1 ORDER BY v.console, v.title') as $row) {
+	INNER JOIN genre g ON v.genre = g.id WHERE v.user_id = ' . $_SESSION['id'] . ' ORDER BY v.console, v.title') as $row) {
 		echo '<tr>';
 		echo '<td>' . $row['console'] . '</td>';
 		echo '<td>' . $row['title'] . '</td>';
@@ -67,6 +97,8 @@ if (isset($_POST)) {
 		echo '<td><button type="submit" value="' . $row['id'] . '"name="delete">Delete Items</button>'; 
 		echo '</tr>';
 	}
-?>
+
+	unset($_SESSION['table']);
+	
 </body>
 </html>
